+++ src/client/package.json
@@ -23,13 +23,13 @@
     "build-css": "node-sass-chokidar --o src/ui/styles/css --source-map true --source-map-contents src/ui/styles/scss",
     "watch-css": "npm run build-css && node-sass-chokidar --o src/ui/styles/css --source-map true --source-map-contents src/ui/styles/scss --watch --recursive",
     "start-js": "react-scripts-ts start",
-    "start": "cross-env REACT_APP_ENV=demo npm-run-all -p watch-css start-js",
+    "start": "npm-run-all -p watch-css start-js",
     "build-js": "react-scripts-ts build",
     "build": "npm-run-all build-css build-js",
     "test": "react-scripts-ts test --env=jsdom",
     "start:local": "cross-env REACT_APP_ENV=local npm-run-all -p watch-css start-js",
     "start:docker": "cross-env REACT_APP_ENV=docker npm-run-all -p watch-css start-js",
-    "start:cloud": "cross-env REACT_APP_ENV=cloud npm-run-all -p watch-css start-js",
+    "start:cloud": "cross-env REACT_APP_ENV=demo npm-run-all -p watch-css start-js",
     "start:openfin:local": "npm-run-all -p openfin start:local",
     "start:openfin:cloud": "npm-run-all -p openfin start:cloud",
     "openfin": "openfin -c config/openfin/local.app.json -l",
@@ -37,6 +37,7 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@types/query-string": "^5.1.0",
     "ag-grid": "15.0.0",
     "ag-grid-react": "15.0.0",
     "animate.css": "3.4.0",
@@ -54,8 +55,8 @@
     "moment": "2.13.0",
     "numeral": "1.5.3",
     "nvd3": "1.8.5",
-    "prettier": "^1.12.1",
     "prop-types": "15.6.0",
+    "query-string": "^6.0.0",
     "react": "^16.3.2",
     "react-dom": "^16.3.2",
     "react-dom-factories": "1.0.2",
