--- src/client/src/index.html
@@ -1,12 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <title>Reactive Trader Cloud</title>
-  <meta charset="utf-8">
-  <link rel="shortcut icon" type="image/x-icon" href="images/favicon.ico"/>
-  <meta name="viewport" content="width=device-width, initial-scale=1"/>
-</head>
-<body>
-  <div id="root"></div>
-</body>
-</html>
