+++ src/client/src/ui/common/popout/openfinPopoutService.tsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 import * as ReactDOM from 'react-dom'
 import OpenFinChrome from '../../shell/OpenFinChrome'
 import PopoutServiceBase from './popoutServiceBase'
-const DockingManager = require('./dockingManager.js')
+const {DockingManager} = require('./dockingManager.js')
 
 const DOCKED_CLASS_NAME = 'docked'
 const BOUNDS_CHANGING_EVENT = 'bounds-changing'
