+++ src/client/src/ui/notification/TradeNotification.tsx
@@ -1,6 +1,7 @@
 import * as classnames from 'classnames'
 import * as React from 'react'
 import { Direction, Trade, TradeStatus } from '../../types'
+import '../styles/css/index.css'
 
 export interface TradeNotificationProps {
   message: Trade
