+++ src/client/src/ui/styles/scss/alertNotification.scss
@@ -1,0 +1,105 @@
+.notification {
+    height: 100vh;
+    background-color: #262a33;
+    user-select: none;
+    position: relative;
+  }
+  
+  .notification__content {
+    position: absolute;
+    top: 5px;
+    left: 5px;
+    right: 5px;
+    bottom: 5px;
+  }
+  
+  .notification__status {
+    position: absolute;
+    right: 0;
+    top: 0;
+    font-size: 11px;
+    color: #c2c5c9;
+  }
+  
+  .notification__summary-items {
+    padding: 0;
+    margin: 0;
+    list-style: none;
+  }
+  
+  .notification__summary-items--rejected {
+    .notification__summary-item {
+      text-decoration: line-through;
+    }
+  }
+  
+  .notification__summary-item {
+    color: #c2c5c9;
+    line-height: 14px;
+  }
+  
+  .notification__summary-item--direction {
+    font-family: BrandonMedium;
+    text-transform: uppercase;
+    font-size: 11px;
+  }
+  
+  .notification__summary-item--notional {
+    color: #f8ab02;
+  }
+  
+  .notification__summary-item--currency {
+    font-size: 12px;
+  }
+  
+  .notification__details-items-container {
+    display: flex;
+    border-top: 1px solid rgba(255, 255, 255, 0.08);
+  }
+  
+  .notification__details-items {
+    flex: 1;
+    padding: 0;
+    margin-top: 1px;
+  }
+  
+  .notification__details-item {
+    list-style: none;
+    margin: 0;
+    padding: 0;
+    line-height: 12px;
+  }
+  
+  .notification__details-item--label {
+    color: #c2c5c9;
+    font-size: 9px;
+    text-transform: uppercase;
+    font-family: BrandonMedium;
+  }
+  
+  .notification__details-item--value {
+    color: white;
+    font-size: 11px;
+    text-transform: uppercase;
+  }
+  
+  .notification__button--dismiss {
+    text-decoration: none;
+    position: absolute;
+    bottom: -5px;
+    right: 0;
+    color: inherit;
+    color: #c2c5c9;
+  }
+  
+  .notification__button--dismiss-icon {
+    transform: rotate(45deg);
+  }
+  
+  .notification__status--done {
+    display: none;
+  }
+  
+  .notification__status--rejected {
+    color: #d32121;
+  }
