+++ src/client/src/ui/styles/scss/index.scss
@@ -20,4 +20,5 @@
 @import 'popoutRegion.scss';
 @import 'spotTileControls.scss';
 @import 'variables';
+@import 'alertNotifcation.scss';
 @import 'shellVariables';
